sync fork

+++ src/GraphXings/Gruppe5Algo/Utils/RationalComputer.java
@@ -4,7 +4,7 @@ import GraphXings.Data.Rational;
 
 public class RationalComputer {
 
-    public static int compareRational(Rational x, Rational y) {
+    public static long compareRational(Rational x, Rational y) {
         var a = x.getP() * y.getQ();
         var b = x.getQ() * y.getP();
         return b - a;
